requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
dlt[snowflake,postgres]>=0.4.0
requests>=2.31.0
apache-airflow-providers-snowflake>=5.0.0
apache-airflow-providers-postgres>=5.0.0
